update music keys to work even if 2 players running
I had cases where I was listening to MPD, but had spotify also open and
needed to pause the music real quick, but Spotify started playing as well
instead. To fix this, now the currently playing player is prioritized, also
it remembers the last player a command was sent to and sends new commands
there as well. This way I can quickly pause and resume either of the
players without worrying about the other starting instead.