update set_wallpaper to separate lock/wallpaper images
i3lock requires the image to be in the PNG format. But I don't
want to convert all the wallpapers to PNG just for the lock
screen. Also, the images have to be resized only for the lock
screen, as feh can manage resizing for the background itself.

Because of all this, the .wallpaper file will be just a copy of
the original image, and the .wallpaper_lock file will be a resized
PNG.